requirements.txt
# Whale Watcher Dependencies

# Async HTTP & WebSockets
aiohttp>=3.9.0
websockets>=12.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Telegram Bot
python-telegram-bot>=20.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
tenacity>=8.2.3
loguru>=0.7.0
pydantic>=2.0.0

# HTTP requests (sync fallback)
requests>=2.31.0

# Date/time handling
python-dateutil>=2.8.0
